+++ Changelog.md
@@ -7,6 +7,9 @@
 ### 1.8.9
 *<small>not released yet</small>*
 
+* **Add Recent Folder List in Folder Browser**   
+The folder browser now has an icon to show and re-select recently opened folders.
+
 * **Fix Shift-Delete Behavior**   
 Fix additional issue with `Shift-Del` operation when text is selected which now properly **cuts** selected text to the clipboard. Previous behavior just removed the selection.